Trait dash_sdk::mock::Mockable

pub trait Mockable: Sized {
    // Provided methods
    fn mock_serialize(&self) -> Option<Vec<u8>> { ... }
    fn mock_deserialize(_data: &[u8]) -> Option<Self> { ... }
}

Mocking support for messages.

This trait should be implemented by any object that can be used in the DAPI.

We use serde_json to serialize/deserialize messages.

Provided Methods

fn mock_serialize(&self) -> Option<Vec<u8>>

Serialize the message to bytes for mocking purposes.

Returns None if the message is not serializable or mocking is disabled.

Panics

Panics on any error.

fn mock_deserialize(_data: &[u8]) -> Option<Self>

Deserialize the message serialized with [mock_serialize()].

Returns None if the message is not serializable or mocking is disabled.

Panics

Panics on any error.

Object Safety

This trait is not object safe.