Enum dash_sdk::platform::DataContract

pub enum DataContract {
    V0(DataContractV0),
}

Understanding Data Contract versioning Data contract versioning is both for the code structure and for serialization.

The code structure is what is used in code to verify documents and is used in memory There is generally only one code structure running at any given time, except in the case we are switching protocol versions.

There can be a lot of serialization versions that are active, and serialization versions should generally always be supported. This is because when we store something as version 1. 10 years down the line when we unserialize this contract it will still be in version 1. Deserialization of a data contract serialized in that version should be translated to the current code structure version.

There are some scenarios to consider,

One such scenario is that the serialization version does not contain enough information for the current code structure version.

Depending on the situation one of the following occurs:

the contract structure can imply missing parts based on default behavior

the contract structure can disable certain features dependant on missing information

the contract might be unusable until it is updated by the owner

Here we use PlatformSerialize, because

Variants

V0(DataContractV0)

Implementations

impl DataContract

pub fn generate_data_contract_id_v0( owner_id: impl AsRef<[u8]>, identity_nonce: u64 ) -> Identifier

Generate data contract id based on owner id and identity nonce

impl DataContract

pub fn try_from_platform_versioned( value: DataContractInSerializationFormat, full_validation: bool, validation_operations: &mut Vec<ProtocolValidationOperation>, platform_version: &PlatformVersion ) -> Result<DataContract, ProtocolError>

impl DataContract

pub fn feature_version(&self) -> u16

impl DataContract

pub fn as_v0(&self) -> Option<&DataContractV0>

pub fn as_v0_mut(&mut self) -> Option<&mut DataContractV0>

pub fn into_v0(self) -> Option<DataContractV0>

pub fn check_version_is_active( protocol_version: u32, data_contract_system_version: u16 ) -> Result<bool, ProtocolError>

pub fn hash( &self, platform_version: &PlatformVersion ) -> Result<Vec<u8>, ProtocolError>