Enum dash_sdk::platform::IdentityPublicKey

pub enum IdentityPublicKey {
    V0(IdentityPublicKeyV0),
}

Variants

V0(IdentityPublicKeyV0)

Implementations

impl IdentityPublicKey

pub fn random_key( id: u32, seed: Option<u64>, platform_version: &PlatformVersion ) -> IdentityPublicKey

pub fn random_keys( first_id: u32, count: u32, seed: Option<u64>, platform_version: &PlatformVersion ) -> Vec<IdentityPublicKey>

pub fn random_authentication_key( key_id: u32, seed: Option<u64>, platform_version: &PlatformVersion ) -> IdentityPublicKey

pub fn random_authentication_keys( first_id: u32, count: u32, seed: Option<u64>, platform_version: &PlatformVersion ) -> Vec<IdentityPublicKey>

pub fn random_authentication_key_with_rng( id: u32, rng: &mut StdRng, used_key_matrix: Option<(u32, &mut Vec<bool>)>, platform_version: &PlatformVersion ) -> Result<IdentityPublicKey, ProtocolError>

Generates a random authentication key based on the platform version.

Parameters

id: The KeyID for the generated key.

rng: A mutable reference to a random number generator of type StdRng.

used_key_matrix: An optional tuple that contains the count of keys that have already been used and a mutable reference to a matrix (or vector) that tracks which keys have been used.

platform_version: The platform version which determines the structure of the identity key.

Returns

Result<Self, ProtocolError>: If successful, returns an instance of Self. In case of an error, it returns a ProtocolError.

Errors

ProtocolError::PublicKeyGenerationError: This error is returned if too many keys have already been created.

ProtocolError::UnknownVersionMismatch: This error is returned if the provided platform version is not recognized.

pub fn random_authentication_key_with_private_key( id: u32, seed: Option<u64>, platform_version: &PlatformVersion ) -> Result<(IdentityPublicKey, Vec<u8>), ProtocolError>

Generates a random authentication key and its corresponding private key based on the platform version.

Parameters

id: The KeyID for the generated key.

seed: A seed that will create a random number generator StdRng.

used_key_matrix: An optional tuple that contains the count of keys that have already been used and a mutable reference to a matrix (or vector) that tracks which keys have been used.

platform_version: The platform version which determines the structure of the identity key.

Returns

Result<(Self, Vec<u8>), ProtocolError>: If successful, returns an instance of Self and the private key as Vec<u8>. In case of an error, it returns a ProtocolError.

Errors

ProtocolError::PublicKeyGenerationError: This error is returned if too many keys have already been created.

ProtocolError::UnknownVersionMismatch: This error is returned if the provided platform version is not recognized.

pub fn random_authentication_key_with_private_key_with_rng( id: u32, rng: &mut StdRng, used_key_matrix: Option<(u32, &mut Vec<bool>)>, platform_version: &PlatformVersion ) -> Result<(IdentityPublicKey, Vec<u8>), ProtocolError>

Generates a random authentication key and its corresponding private key based on the platform version.

Parameters

id: The KeyID for the generated key.

rng: A mutable reference to a random number generator of type StdRng.

used_key_matrix: An optional tuple that contains the count of keys that have already been used and a mutable reference to a matrix (or vector) that tracks which keys have been used.

platform_version: The platform version which determines the structure of the identity key.

Returns

Result<(Self, Vec<u8>), ProtocolError>: If successful, returns an instance of Self and the private key as Vec<u8>. In case of an error, it returns a ProtocolError.

Errors

ProtocolError::PublicKeyGenerationError: This error is returned if too many keys have already been created.

ProtocolError::UnknownVersionMismatch: This error is returned if the provided platform version is not recognized.

pub fn random_key_with_rng( id: u32, rng: &mut StdRng, used_key_matrix: Option<(u32, &mut Vec<bool>)>, platform_version: &PlatformVersion ) -> Result<IdentityPublicKey, ProtocolError>

Generates a random key based on the platform version.

Parameters

id: The KeyID for the generated key.

rng: A mutable reference to a random number generator of type StdRng.

used_key_matrix: An optional tuple that contains the count of keys that have already been used and a mutable reference to a matrix (or vector) that tracks which keys have been used.

platform_version: The platform version which determines the structure of the identity key.

Returns

Result<Self, ProtocolError>: If successful, returns an instance of Self. In case of an error, it returns a ProtocolError.

Errors

ProtocolError::PublicKeyGenerationError: This error is returned if too many keys have already been created.

ProtocolError::UnknownVersionMismatch: This error is returned if the provided platform version is not recognized.

pub fn random_key_with_known_attributes( id: u32, rng: &mut StdRng, purpose: Purpose, security_level: SecurityLevel, key_type: KeyType, contract_bounds: Option<ContractBounds>, platform_version: &PlatformVersion ) -> Result<(IdentityPublicKey, Vec<u8>), ProtocolError>

Generates a random key based on the platform version.

Parameters

id: The KeyID for the generated key.

rng: A mutable reference to a random number generator of type StdRng.

used_key_matrix: An optional tuple that contains the count of keys that have already been used and a mutable reference to a matrix (or vector) that tracks which keys have been used.

platform_version: The platform version which determines the structure of the identity key.

Returns

Result<Self, ProtocolError>: If successful, returns an instance of Self. In case of an error, it returns a ProtocolError.

Errors

ProtocolError::PublicKeyGenerationError: This error is returned if too many keys have already been created.

ProtocolError::UnknownVersionMismatch: This error is returned if the provided platform version is not recognized.

pub fn random_ecdsa_master_authentication_key_with_rng( id: u32, rng: &mut StdRng, platform_version: &PlatformVersion ) -> Result<(IdentityPublicKey, Vec<u8>), ProtocolError>

Generates a random ECDSA master authentication public key along with its corresponding private key.

This method constructs a random ECDSA (using the secp256k1 curve) master authentication public key and returns both the public key and its corresponding private key.

Parameters

id: The KeyID for the generated key.

rng: A mutable reference to the random number generator.

Returns

(Self, Vec<u8>): A tuple where the first element is an instance of the IdentityPublicKey struct, and the second element is the corresponding private key.

pub fn random_ecdsa_master_authentication_key( id: u32, seed: Option<u64>, platform_version: &PlatformVersion ) -> Result<(IdentityPublicKey, Vec<u8>), ProtocolError>

Generates a random ECDSA master-level authentication public key along with its corresponding private key.

This method constructs a random ECDSA (using the secp256k1 curve) high-level authentication public key and returns both the public key and its corresponding private key.

Parameters

id: The KeyID for the generated key.

seed: A seed that will create a random number generator StdRng.

Returns

(Self, Vec<u8>): A tuple where the first element is an instance of the IdentityPublicKey struct, and the second element is the corresponding private key.

pub fn random_ecdsa_critical_level_authentication_key( id: u32, seed: Option<u64>, platform_version: &PlatformVersion ) -> Result<(IdentityPublicKey, Vec<u8>), ProtocolError>

Generates a random ECDSA critical-level authentication public key along with its corresponding private key.

This method constructs a random ECDSA (using the secp256k1 curve) high-level authentication public key and returns both the public key and its corresponding private key.

Parameters

id: The KeyID for the generated key.

seed: A seed that will create a random number generator StdRng.

Returns

(Self, Vec<u8>): A tuple where the first element is an instance of the IdentityPublicKey struct, and the second element is the corresponding private key.

pub fn random_ecdsa_critical_level_authentication_key_with_rng( id: u32, rng: &mut StdRng, platform_version: &PlatformVersion ) -> Result<(IdentityPublicKey, Vec<u8>), ProtocolError>

Generates a random ECDSA high-level authentication public key along with its corresponding private key.

This method constructs a random ECDSA (using the secp256k1 curve) high-level authentication public key and returns both the public key and its corresponding private key.

Parameters

id: The KeyID for the generated key.

rng: A mutable reference to the random number generator.

Returns

(Self, Vec<u8>): A tuple where the first element is an instance of the IdentityPublicKey struct, and the second element is the corresponding private key.

pub fn random_ecdsa_high_level_authentication_key( id: u32, seed: Option<u64>, platform_version: &PlatformVersion ) -> Result<(IdentityPublicKey, Vec<u8>), ProtocolError>

Generates a random ECDSA high-level authentication public key along with its corresponding private key.

This method constructs a random ECDSA (using the secp256k1 curve) high-level authentication public key and returns both the public key and its corresponding private key.

Parameters

id: The KeyID for the generated key.

seed: A seed that will create a random number generator StdRng.

Returns

(Self, Vec<u8>): A tuple where the first element is an instance of the IdentityPublicKey struct, and the second element is the corresponding private key.

pub fn random_ecdsa_high_level_authentication_key_with_rng( id: u32, rng: &mut StdRng, platform_version: &PlatformVersion ) -> Result<(IdentityPublicKey, Vec<u8>), ProtocolError>

Generates a random ECDSA high-level authentication public key along with its corresponding private key.

This method constructs a random ECDSA (using the secp256k1 curve) high-level authentication public key and returns both the public key and its corresponding private key.

Parameters

id: The KeyID for the generated key.

rng: A mutable reference to the random number generator.

Returns

(Self, Vec<u8>): A tuple where the first element is an instance of the IdentityPublicKey struct, and the second element is the corresponding private key.

pub fn random_authentication_keys_with_rng( key_count: u32, rng: &mut StdRng, platform_version: &PlatformVersion ) -> Result<Vec<IdentityPublicKey>, ProtocolError>

pub fn random_authentication_keys_with_private_keys_with_rng( start_id: u32, key_count: u32, rng: &mut StdRng, platform_version: &PlatformVersion ) -> Result<Vec<(IdentityPublicKey, Vec<u8>)>, ProtocolError>

pub fn main_keys_with_random_authentication_keys_with_private_keys_with_rng( key_count: u32, rng: &mut StdRng, platform_version: &PlatformVersion ) -> Result<Vec<(IdentityPublicKey, Vec<u8>)>, ProtocolError>

pub fn random_unique_keys_with_rng( key_count: u32, rng: &mut StdRng, platform_version: &PlatformVersion ) -> Result<Vec<IdentityPublicKey>, ProtocolError>

pub fn random_keys_with_rng( key_count: u32, rng: &mut StdRng, platform_version: &PlatformVersion ) -> Vec<IdentityPublicKey>

impl IdentityPublicKey

pub fn is_master(&self) -> bool

Checks if public key security level is MASTER

pub fn max_possible_size_key( id: u32, platform_version: &PlatformVersion ) -> Result<IdentityPublicKey, ProtocolError>

Generates an identity public key with the maximum possible size based on the platform version.

This method constructs a key of the largest possible size for the given platform version. This can be useful for stress testing or benchmarking purposes.

Parameters

id: The KeyID for the generated key.

platform_version: The platform version which determines the structure of the identity key.

Returns

Self: An instance of the IdentityPublicKey struct.

pub fn default_versioned( platform_version: &PlatformVersion ) -> Result<IdentityPublicKey, ProtocolError>