dash_sdk::platform
Module proto

Modules

get_consensus_params_request

Nested message and enum types in GetConsensusParamsRequest.

get_consensus_params_response

Nested message and enum types in GetConsensusParamsResponse.

get_contested_resource_identity_votes_request

Nested message and enum types in GetContestedResourceIdentityVotesRequest.

get_contested_resource_identity_votes_response

Nested message and enum types in GetContestedResourceIdentityVotesResponse.

get_contested_resource_vote_state_request

Nested message and enum types in GetContestedResourceVoteStateRequest.

get_contested_resource_vote_state_response

Nested message and enum types in GetContestedResourceVoteStateResponse.

get_contested_resource_voters_for_identity_request

Nested message and enum types in GetContestedResourceVotersForIdentityRequest.

get_contested_resource_voters_for_identity_response

Nested message and enum types in GetContestedResourceVotersForIdentityResponse.

get_contested_resources_request

Nested message and enum types in GetContestedResourcesRequest.

get_contested_resources_response

Nested message and enum types in GetContestedResourcesResponse.

get_current_quorums_info_request

Nested message and enum types in GetCurrentQuorumsInfoRequest.

get_current_quorums_info_response

Nested message and enum types in GetCurrentQuorumsInfoResponse.

get_data_contract_history_request

Nested message and enum types in GetDataContractHistoryRequest.

get_data_contract_history_response

Nested message and enum types in GetDataContractHistoryResponse.

get_data_contract_request

Nested message and enum types in GetDataContractRequest.

get_data_contract_response

Nested message and enum types in GetDataContractResponse.

get_data_contracts_request

Nested message and enum types in GetDataContractsRequest.

get_data_contracts_response

Nested message and enum types in GetDataContractsResponse.

get_documents_request

Nested message and enum types in GetDocumentsRequest.

get_documents_response

Nested message and enum types in GetDocumentsResponse.

get_epochs_info_request

Nested message and enum types in GetEpochsInfoRequest.

get_epochs_info_response

Nested message and enum types in GetEpochsInfoResponse.

get_evonodes_proposed_epoch_blocks_by_ids_request

Nested message and enum types in GetEvonodesProposedEpochBlocksByIdsRequest.

get_evonodes_proposed_epoch_blocks_by_range_request

Nested message and enum types in GetEvonodesProposedEpochBlocksByRangeRequest.

get_evonodes_proposed_epoch_blocks_response

Nested message and enum types in GetEvonodesProposedEpochBlocksResponse.

get_identities_balances_request

Nested message and enum types in GetIdentitiesBalancesRequest.

get_identities_balances_response

Nested message and enum types in GetIdentitiesBalancesResponse.

get_identities_contract_keys_request

Nested message and enum types in GetIdentitiesContractKeysRequest.

get_identities_contract_keys_response

Nested message and enum types in GetIdentitiesContractKeysResponse.

get_identity_balance_and_revision_request

Nested message and enum types in GetIdentityBalanceAndRevisionRequest.

get_identity_balance_and_revision_response

Nested message and enum types in GetIdentityBalanceAndRevisionResponse.

get_identity_balance_request

Nested message and enum types in GetIdentityBalanceRequest.

get_identity_balance_response

Nested message and enum types in GetIdentityBalanceResponse.

get_identity_by_public_key_hash_request

Nested message and enum types in GetIdentityByPublicKeyHashRequest.

get_identity_by_public_key_hash_response

Nested message and enum types in GetIdentityByPublicKeyHashResponse.

get_identity_contract_nonce_request

Nested message and enum types in GetIdentityContractNonceRequest.

get_identity_contract_nonce_response

Nested message and enum types in GetIdentityContractNonceResponse.

get_identity_keys_request

Nested message and enum types in GetIdentityKeysRequest.

get_identity_keys_response

Nested message and enum types in GetIdentityKeysResponse.

get_identity_nonce_request

Nested message and enum types in GetIdentityNonceRequest.

get_identity_nonce_response

Nested message and enum types in GetIdentityNonceResponse.

get_identity_request

Nested message and enum types in GetIdentityRequest.

get_identity_response

Nested message and enum types in GetIdentityResponse.

get_path_elements_request

Nested message and enum types in GetPathElementsRequest.

get_path_elements_response

Nested message and enum types in GetPathElementsResponse.

get_prefunded_specialized_balance_request

Nested message and enum types in GetPrefundedSpecializedBalanceRequest.

get_prefunded_specialized_balance_response

Nested message and enum types in GetPrefundedSpecializedBalanceResponse.

get_proofs_request

Nested message and enum types in GetProofsRequest.

get_proofs_response

Nested message and enum types in GetProofsResponse.

get_protocol_version_upgrade_state_request

Nested message and enum types in GetProtocolVersionUpgradeStateRequest.

get_protocol_version_upgrade_state_response

Nested message and enum types in GetProtocolVersionUpgradeStateResponse.

get_protocol_version_upgrade_vote_status_request

Nested message and enum types in GetProtocolVersionUpgradeVoteStatusRequest.

get_protocol_version_upgrade_vote_status_response

Nested message and enum types in GetProtocolVersionUpgradeVoteStatusResponse.

get_status_request

Nested message and enum types in GetStatusRequest.

get_status_response

Nested message and enum types in GetStatusResponse.

get_total_credits_in_platform_request

Nested message and enum types in GetTotalCreditsInPlatformRequest.

get_total_credits_in_platform_response

Nested message and enum types in GetTotalCreditsInPlatformResponse.

get_vote_polls_by_end_date_request

Nested message and enum types in GetVotePollsByEndDateRequest.

get_vote_polls_by_end_date_response

Nested message and enum types in GetVotePollsByEndDateResponse.

key_request_type

Nested message and enum types in KeyRequestType.

platform_client

Generated client implementations.

security_level_map

Nested message and enum types in SecurityLevelMap.

wait_for_state_transition_result_request

Nested message and enum types in WaitForStateTransitionResultRequest.

wait_for_state_transition_result_response

Nested message and enum types in WaitForStateTransitionResultResponse.

Structs

AllKeys

AllKeys is an empty message used to signify a request for all keys

BroadcastStateTransitionRequest

BroadcastStateTransitionResponse

GetConsensusParamsRequest

GetConsensusParamsResponse

GetContestedResourceIdentityVotesRequest

How did an identity vote?

GetContestedResourceIdentityVotesResponse

GetContestedResourceVoteStateRequest

What’s the state of a contested resource vote? (ie who is winning?)

GetContestedResourceVoteStateResponse

GetContestedResourceVotersForIdentityRequest

Who voted for a contested resource to go to a specific identity?

GetContestedResourceVotersForIdentityResponse

GetContestedResourcesRequest

GetContestedResourcesResponse

GetCurrentQuorumsInfoRequest

GetCurrentQuorumsInfoResponse

GetDataContractHistoryRequest

GetDataContractHistoryResponse

GetDataContractRequest

GetDataContractResponse

GetDataContractsRequest

GetDataContractsResponse

GetDocumentsRequest

GetDocumentsResponse

GetEpochsInfoRequest

GetEpochsInfoResponse

GetEvonodesProposedEpochBlocksByIdsRequest

GetEvonodesProposedEpochBlocksByRangeRequest

GetEvonodesProposedEpochBlocksResponse

GetIdentitiesBalancesRequest

GetIdentitiesBalancesResponse

GetIdentitiesContractKeysRequest

GetIdentitiesContractKeysResponse

GetIdentityBalanceAndRevisionRequest

GetIdentityBalanceAndRevisionResponse

GetIdentityBalanceRequest

GetIdentityBalanceResponse

GetIdentityByPublicKeyHashRequest

GetIdentityByPublicKeyHashResponse

GetIdentityContractNonceRequest

GetIdentityContractNonceResponse

GetIdentityKeysRequest

GetIdentityKeysResponse

GetIdentityNonceRequest

GetIdentityNonceResponse

GetIdentityRequest

GetIdentityResponse

GetPathElementsRequest

GetPathElementsResponse

GetPrefundedSpecializedBalanceRequest

GetPrefundedSpecializedBalanceResponse

GetProofsRequest

GetProofsResponse

GetProtocolVersionUpgradeStateRequest

GetProtocolVersionUpgradeStateResponse

GetProtocolVersionUpgradeVoteStatusRequest

GetProtocolVersionUpgradeVoteStatusResponse

GetStatusRequest

GetStatusResponse

GetTotalCreditsInPlatformRequest

GetTotalCreditsInPlatformResponse

GetVotePollsByEndDateRequest

GetVotePollsByEndDateResponse

KeyRequestType

Proof

Proof message includes cryptographic proofs for validating responses

ResponseMetadata

ResponseMetadata provides metadata about the blockchain state at the time of response

SearchKey

SearchKey represents a request to search for keys based on specific criteria

SecurityLevelMap

SecurityLevelMap maps security levels to a request type for key retrieval

SpecificKeys

SpecificKeys is used to request specific keys by their IDs

StateTransitionBroadcastError

WaitForStateTransitionResultRequest

WaitForStateTransitionResultResponse

Enums

KeyPurpose