Trait dash_sdk::platform::FetchMany

pub trait FetchMany<K: Ord>
where
    Self: Sized,
    BTreeMap<K, Option<Self>>: MockResponse + FromProof<Self::Request, Request = Self::Request, Response = <Self::Request as TransportRequest>::Response> + Sync,
{
    type Request: TransportRequest + Into<<BTreeMap<K, Option<Self>> as FromProof<<Self as FetchMany<K>>::Request>>::Request>;

    // Provided methods
    fn fetch_many<'life0, 'async_trait, Q>(
        sdk: &'life0 Sdk,
        query: Q
    ) -> Pin<Box<dyn Future<Output = Result<RetrievedObjects<K, Self>, Error>> + Send + 'async_trait>>
       where Q: 'async_trait + Query<<Self as FetchMany<K>>::Request>,
             Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn fetch_by_identifiers<'life0, 'async_trait, I>(
        sdk: &'life0 Sdk,
        identifiers: I
    ) -> Pin<Box<dyn Future<Output = Result<RetrievedObjects<K, Self>, Error>> + Send + 'async_trait>>
       where Vec<Identifier>: Query<<Self as FetchMany<K>>::Request>,
             I: 'async_trait + IntoIterator<Item = Identifier> + Send,
             Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn fetch_many_with_limit<'life0, 'async_trait, Q>(
        sdk: &'life0 Sdk,
        query: Q,
        limit: u32
    ) -> Pin<Box<dyn Future<Output = Result<RetrievedObjects<K, Self>, Error>> + Send + 'async_trait>>
       where LimitQuery<Q>: Query<<Self as FetchMany<K>>::Request>,
             Q: 'async_trait + Query<<Self as FetchMany<K>>::Request>,
             Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
}

Fetch multiple objects from the Platform.

To fetch multiple objects from the platform, you need to define some query (criteria that fetched objects must match) and use FetchMany::fetch_many() for your object type.

You can also use convenience methods:

[FetchMany::fetch_many_by_identifiers()] - to fetch multiple objects by their identifiers,

FetchMany::fetch_many_with_limit() - to fetch not more than limit objects.

Generic Parameters

K: The type of the key used to index the object

Example

An example use case is to fetch multiple Data Contracts by their Identifiers. As [&[Identifier]] implements Query for this type of requests, you need to:

define Identifiers of data contracts to fetch,

create a query by grouping identifiers in a collection, like a Vec or a slice,

call DataContract::fetch_many() with the query and an instance of Sdk.

use dash_sdk::{Sdk, platform::{Query, Identifier, FetchMany, DataContract}};

let sdk = Sdk::new_mock();

let id1 = Identifier::new(SOME_IDENTIFIER_1);
let id2 = Identifier::new(SOME_IDENTIFIER_2);

let query = vec![id1, id2];

let data_contract = DataContract::fetch_many(&sdk, query);

Required Associated Types

type Request: TransportRequest + Into<<BTreeMap<K, Option<Self>> as FromProof<<Self as FetchMany<K>>::Request>>::Request>

Type of request used to fetch multiple objects from the platform.

Most likely, one of the types defined in dapi_grpc::platform::v0.

This type must implement [TransportRequest] and MockRequest.

Provided Methods

fn fetch_many<'life0, 'async_trait, Q>( sdk: &'life0 Sdk, query: Q ) -> Pin<Box<dyn Future<Output = Result<RetrievedObjects<K, Self>, Error>> + Send + 'async_trait>>
where Q: 'async_trait + Query<<Self as FetchMany<K>>::Request>, Self: Send + 'async_trait, 'life0: 'async_trait,

Fetch (or search) multiple objects on the Dash Platform

FetchMany::fetch_many() is an asynchronous method that fetches multiple objects from the Dash Platform.

Note that this method might introduce some predefined limit on the number of objects returned. If you need to specify the limit yourself, use FetchMany::fetch_many_with_limit() or LimitQuery instead.

Per-object type documentation

See documentation of FetchMany trait implementations for each object type for more details, including list of supported queries.

Generic Parameters

K: The type of the key used to index the object in the returned collection

Parameters

sdk: An instance of Sdk.

query: A query parameter implementing Query to specify the data to be retrieved.

Returns

Returns a Result containing either :

list of objects matching the Query indexed by a key type K, where an item can be None of the object was not found for provided key

Error.

Note that behavior when no items are found can be either empty collection or collection containing None values.

Usage

See tests/fetch/document.rs for a full example.

Error Handling

Any errors encountered during the execution are returned as Error instances.

fn fetch_by_identifiers<'life0, 'async_trait, I>( sdk: &'life0 Sdk, identifiers: I ) -> Pin<Box<dyn Future<Output = Result<RetrievedObjects<K, Self>, Error>> + Send + 'async_trait>>
where Vec<Identifier>: Query<<Self as FetchMany<K>>::Request>, I: 'async_trait + IntoIterator<Item = Identifier> + Send, Self: Send + 'async_trait, 'life0: 'async_trait,

Fetch multiple objects from the Platform by their identifiers.

Convenience method to fetch multiple objects by their identifiers. See FetchMany and FetchMany::fetch_many() for more detailed documentation.

Parameters

sdk: An instance of Sdk.

identifiers: A collection of Identifiers to fetch.

Requirements

Vec<Identifier> must implement Query for Self::Request.

fn fetch_many_with_limit<'life0, 'async_trait, Q>( sdk: &'life0 Sdk, query: Q, limit: u32 ) -> Pin<Box<dyn Future<Output = Result<RetrievedObjects<K, Self>, Error>> + Send + 'async_trait>>
where LimitQuery<Q>: Query<<Self as FetchMany<K>>::Request>, Q: 'async_trait + Query<<Self as FetchMany<K>>::Request>, Self: Send + 'async_trait, 'life0: 'async_trait,

Fetch multiple objects from the Platform with limit.

Fetches up to limit objects matching the query.
See FetchMany and FetchMany::fetch_many() for more detailed documentation.

Parameters

sdk: An instance of Sdk.

query: A query parameter implementing Query to specify the data to be retrieved.

limit: Maximum number of objects to fetch.

Object Safety

This trait is not object safe.