dash_sdk::platform
Trait FetchUnproved

pub trait FetchUnproved
where
    Self: Sized + Debug + MockResponse,
{
    type Request: TransportRequest;

    // Provided methods
    fn fetch_unproved<'life0, 'async_trait, Q>(
        sdk: &'life0 Sdk,
        query: Q,
    ) -> Pin<Box<dyn Future<Output = Result<Option<Self>, Error>> + Send + 'async_trait>>
       where Self: FromUnproved<<Self as FetchUnproved>::Request, Request = <Self as FetchUnproved>::Request, Response = <<Self as FetchUnproved>::Request as TransportRequest>::Response> + Send + 'async_trait,
             Q: 'async_trait + Query<<Self as FetchUnproved>::Request>,
             'life0: 'async_trait { ... }
    fn fetch_unproved_with_settings<'life0, 'async_trait, Q>(
        sdk: &'life0 Sdk,
        query: Q,
        settings: RequestSettings,
    ) -> Pin<Box<dyn Future<Output = Result<(Option<Self>, ResponseMetadata), Error>> + Send + 'async_trait>>
       where Self: FromUnproved<<Self as FetchUnproved>::Request, Request = <Self as FetchUnproved>::Request, Response = <<Self as FetchUnproved>::Request as TransportRequest>::Response> + Send + 'async_trait,
             Q: 'async_trait + Query<<Self as FetchUnproved>::Request>,
             'life0: 'async_trait { ... }
}

Required Associated Types

type Request: TransportRequest

Type of request used to fetch data from Platform.

Provided Methods

fn fetch_unproved<'life0, 'async_trait, Q>( sdk: &'life0 Sdk, query: Q, ) -> Pin<Box<dyn Future<Output = Result<Option<Self>, Error>> + Send + 'async_trait>>
where Self: FromUnproved<<Self as FetchUnproved>::Request, Request = <Self as FetchUnproved>::Request, Response = <<Self as FetchUnproved>::Request as TransportRequest>::Response> + Send + 'async_trait, Q: 'async_trait + Query<<Self as FetchUnproved>::Request>, 'life0: 'async_trait,

Fetch unproved data from the Platform.

Parameters

sdk: An instance of Sdk.

query: Query used to fetch data from the Platform.

Returns

Returns:

Ok(Some(Self)) when object is found.

Ok(None) when object is not found.

Err(Error) when an error occurs.

fn fetch_unproved_with_settings<'life0, 'async_trait, Q>( sdk: &'life0 Sdk, query: Q, settings: RequestSettings, ) -> Pin<Box<dyn Future<Output = Result<(Option<Self>, ResponseMetadata), Error>> + Send + 'async_trait>>
where Self: FromUnproved<<Self as FetchUnproved>::Request, Request = <Self as FetchUnproved>::Request, Response = <<Self as FetchUnproved>::Request as TransportRequest>::Response> + Send + 'async_trait, Q: 'async_trait + Query<<Self as FetchUnproved>::Request>, 'life0: 'async_trait,

Fetch unproved data from the Platform with custom settings.

Parameters

sdk: An instance of Sdk.

query: Query used to fetch data from the Platform.

settings: Request settings for the connection to Platform.

Returns

Ok(Some(Self)) when object is found.

Ok(None) when object is not found.

Err(Error) when an error occurs.

Object Safety

This trait is not object safe.