Trait dash_sdk::platform::Query

pub trait Query<T: TransportRequest + Mockable>: Send + Debug + Clone {
    // Required method
    fn query(self, prove: bool) -> Result<T, Error>;
}

Trait implemented by objects that can be used as queries.

Query trait is used to specify criteria for fetching data from the platform. It can be used to specify individual objects as well as search criteria for fetching multiple objects from the platform.

Some examples of queries include:

Identifier - fetches an object by its identifier; implemented for Identity, DataContract and Document.

DocumentQuery - fetches Document based on search conditions; see query syntax documentation for more details.

Example

To fetch individual Identity object by its Identifier, you just need to create it and use Fetch or FetchMany trait:

use dash_sdk::{Sdk, platform::{Query, Identifier, Fetch, Identity}};

let sdk = Sdk::new_mock();
let query = Identifier::new(SOME_IDENTIFIER);
let identity = Identity::fetch(&sdk, query);

As Identifier implements Query, the query variable in the code above can be used as a parameter for Fetch::fetch() and FetchMany::fetch_many() methods.

Required Methods

fn query(self, prove: bool) -> Result<T, Error>

Converts the current instance into an instance of the TransportRequest type.

This method takes ownership of the instance upon which it’s called (hence self), and attempts to perform the conversion.

Arguments

prove - Whether to include proofs in the response. Only true is supported at the moment.

Returns

On success, this method yields an instance of the TransportRequest type (T). On failure, it yields an Error.

Error Handling

This method propagates any errors encountered during the conversion process. These are returned as Error instances.

Object Safety

This trait is not object safe.