Module dash_sdk::platform::transition

State transitions used to put changed objects to the Dash Platform.

Modules

broadcast

broadcast_request

Broadcast trait representing the action of broadcasting a new identity state transition to the platform.

purchase_document

put_contract

put_document

put_identity

put_settings

top_up_identity

transfer_document

update_price_of_document

withdraw_from_identity

Structs

TxId

State transition identifier

Enums

TransitionContext