dash_sdk::platform::transition::transfer
Trait TransferToIdentity

pub trait TransferToIdentity: Waitable {
    // Required method
    fn transfer_credits<'life0, 'life1, 'life2, 'async_trait, S>(
        &'life0 self,
        sdk: &'life1 Sdk,
        to_identity_id: Identifier,
        amount: u64,
        signing_transfer_key_to_use: Option<&'life2 IdentityPublicKey>,
        signer: S,
        settings: Option<PutSettings>,
    ) -> Pin<Box<dyn Future<Output = Result<u64, Error>> + Send + 'async_trait>>
       where S: 'async_trait + Signer + Send,
             Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
}

Required Methods

fn transfer_credits<'life0, 'life1, 'life2, 'async_trait, S>( &'life0 self, sdk: &'life1 Sdk, to_identity_id: Identifier, amount: u64, signing_transfer_key_to_use: Option<&'life2 IdentityPublicKey>, signer: S, settings: Option<PutSettings>, ) -> Pin<Box<dyn Future<Output = Result<u64, Error>> + Send + 'async_trait>>
where S: 'async_trait + Signer + Send, Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Function to transfer credits from an identity to another identity. Returns the final identity balance.

If signing_transfer_key_to_use is not set, we will try to use one in the signer that is available for the transfer.

This method will resolve once the state transition is executed.

Returns

Final balance of the identity after the transfer.

Object Safety

This trait is not object safe.