dash_sdk::platform::transition::update_price_of_document
Trait UpdatePriceOfDocument

pub trait UpdatePriceOfDocument<S: Signer>: Waitable {
    // Required methods
    fn update_price_of_document<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        price: Credits,
        sdk: &'life1 Sdk,
        document_type: DocumentType,
        identity_public_key: IdentityPublicKey,
        signer: &'life2 S,
        settings: Option<PutSettings>,
    ) -> Pin<Box<dyn Future<Output = Result<StateTransition, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn update_price_of_document_and_wait_for_response<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        price: Credits,
        sdk: &'life1 Sdk,
        document_type: DocumentType,
        identity_public_key: IdentityPublicKey,
        signer: &'life2 S,
        settings: Option<PutSettings>,
    ) -> Pin<Box<dyn Future<Output = Result<Document, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
}

A trait for updating the price of a document on Platform

Required Methods

fn update_price_of_document<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, price: Credits, sdk: &'life1 Sdk, document_type: DocumentType, identity_public_key: IdentityPublicKey, signer: &'life2 S, settings: Option<PutSettings>, ) -> Pin<Box<dyn Future<Output = Result<StateTransition, Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Updates the price of a document on platform Setting settings to None sets default connection behavior

fn update_price_of_document_and_wait_for_response<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, price: Credits, sdk: &'life1 Sdk, document_type: DocumentType, identity_public_key: IdentityPublicKey, signer: &'life2 S, settings: Option<PutSettings>, ) -> Pin<Box<dyn Future<Output = Result<Document, Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Updates the price of a document on platform and waits for the response

Object Safety

This trait is not object safe.