dash_sdk::platform::transition::withdraw_from_identity
Trait WithdrawFromIdentity

pub trait WithdrawFromIdentity {
    // Required method
    fn withdraw<'life0, 'life1, 'life2, 'async_trait, S>(
        &'life0 self,
        sdk: &'life1 Sdk,
        address: Option<Address>,
        amount: u64,
        core_fee_per_byte: Option<u32>,
        signing_withdrawal_key_to_use: Option<&'life2 IdentityPublicKey>,
        signer: S,
        settings: Option<PutSettings>,
    ) -> Pin<Box<dyn Future<Output = Result<u64, Error>> + Send + 'async_trait>>
       where S: 'async_trait + Signer + Send,
             Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
}

Required Methods

fn withdraw<'life0, 'life1, 'life2, 'async_trait, S>( &'life0 self, sdk: &'life1 Sdk, address: Option<Address>, amount: u64, core_fee_per_byte: Option<u32>, signing_withdrawal_key_to_use: Option<&'life2 IdentityPublicKey>, signer: S, settings: Option<PutSettings>, ) -> Pin<Box<dyn Future<Output = Result<u64, Error>> + Send + 'async_trait>>
where S: 'async_trait + Signer + Send, Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Function to withdraw credits from an identity. Returns the final identity balance. If signing_withdrawal_key_to_use is not set, we will try to use one in the signer that is available for withdrawal

Object Safety

This trait is not object safe.