dash_sdk
Module query_types

Object types that can be retrieved from proofs.

Some DAPI requests return response types that are not defined in the Dash Platform Protocol, like GetIdentityBalanceRequest which returns u64. In this case, the FromProof trait is implemented for dedicated object type defined in this module.

Structs

Chain

Layer 2 blockchain information

Contenders

Multiple contenders for a vote resolution.

ContestedResource

Contested resource values.

ContestedResources

Contested resources

ContestedVote

A contested vote for querying

CurrentQuorumsInfo

Represents the current state of quorums in the platform.

DriveProtocol

Drive protocol versions.

ElementFetchRequestItem

The item of an element fetch request

EvoNodeStatus

The status of an EvoNode.

IdentityContractNonceFetcher

An identity contract nonce

IdentityNonceFetcher

An identity nonce

IndexMap

A hash table where the iteration order of the key-value pairs is independent of the hash values of the keys.

KeysInPath

Keys in a Path

MasternodeProtocolVote

Vote of a masternode for a protocol version.

Network

Node networking information.

NoParamQuery

A query with no parameters

Node

Evo node identification information.

PrefundedSpecializedBalance

Prefunded specialized balance.

ProposerBlockCountById

A block count struct

ProposerBlockCountByRange

A block count struct

ProposerBlockCounts

Proposer block counts

Protocol

Information about protocol-level versions used by the node

Software

Information about software components versions.

StateSync

Information about state synchronization progress.

TenderdashProtocol

Tenderdash protocol versions.

Time

Information about current time used by the node.

TotalCreditsInPlatform

The total credits on Platform.

Version

Information about protocol and software components versions.

VotePollsGroupedByTimestamp

Contested document resource vote polls grouped by timestamp.

Voter

Identifier of a single voter

Voters

Multiple voters.

Type Aliases

DataContractHistory

History of a data contract.

DataContracts

Multiple data contracts.

Documents

Collection of documents.

Elements

Multiple grovedb elements.

ExtendedEpochInfos

Collection of epoch information

IdentityBalance

Identity balance.

IdentityBalanceAndRevision

Identity balance and revision of the identity.

IdentityBalances

Collection of balances.

IdentityPublicKeys

Public keys belonging to some identity.

MasternodeProtocolVotes

Information about protocol version voted by each node.

ProtocolVersionUpgrades

Results of protocol version upgrade voting.

ProtocolVersionVoteCount

Number of votes for a protocol version upgrade.

ResourceVotesByIdentity

Votes casted by some identity.

RetrievedObjects

A data structure that holds a set of objects of a generic type O, indexed by a key of type K.

RetrievedValues

A data structure that holds a set of values of a generic type I, indexed by a key of type K.