dash_sdk
Module platform

Platform DAPI requests.

Re-exports

pub use dpp;

pub use rs_dapi_client as dapi;

Modules

block_info_from_metadata

fetch_current_no_parameters

proto

transition

State transitions used to put changed objects to the Dash Platform.

types

Type-specific implementation for various dpp object types to make queries more convenient and intuitive.

Structs

DocumentQuery

Request that is used to query documents from the Dash Platform.

DriveDocumentQuery

Drive query struct

Identifier

LimitQuery

Wrapper around query that allows to specify limit.

MockContextProvider

Mock ContextProvider that can read quorum keys from files.

QueryStartInfo

Enums

DataContract

Understanding Data Contract versioning Data contract versioning is both for the code structure and for serialization.

Document

Identity

The identity is not stored inside of drive, because of this, the serialization is mainly for transport, the serialization of the identity will include the version, so no passthrough or untagged is needed here

IdentityPublicKey

Constants

DEFAULT_EPOCH_QUERY_LIMIT

Default limit of epoch records returned by Platform.

Traits

ContextProvider

Interface between the Sdk and state of the application.

Fetch

Trait implemented by objects that can be fetched from Platform.

FetchMany

Fetch multiple objects from Platform.

FetchUnproved

Query

Trait implemented by objects that can be used as queries.

Type Aliases

Revision