dash_sdk::platform
Enum IdentityPublicKey

pub enum IdentityPublicKey {
    V0(IdentityPublicKeyV0),
}

Variants

V0(IdentityPublicKeyV0)

Implementations

impl IdentityPublicKey

pub fn random_key( id: u32, seed: Option<u64>, platform_version: &PlatformVersion, ) -> IdentityPublicKey

pub fn random_keys( first_id: u32, count: u32, seed: Option<u64>, platform_version: &PlatformVersion, ) -> Vec<IdentityPublicKey>

pub fn random_authentication_key( key_id: u32, seed: Option<u64>, platform_version: &PlatformVersion, ) -> IdentityPublicKey

pub fn random_authentication_keys( first_id: u32, count: u32, seed: Option<u64>, platform_version: &PlatformVersion, ) -> Vec<IdentityPublicKey>

pub fn random_authentication_key_with_rng( id: u32, rng: &mut StdRng, used_key_matrix: Option<(u32, &mut Vec<bool>)>, platform_version: &PlatformVersion, ) -> Result<IdentityPublicKey, ProtocolError>

Generates a random authentication key based on the platform version.

Parameters

id: The KeyID for the generated key.

rng: A mutable reference to a random number generator of type StdRng.

used_key_matrix: An optional tuple that contains the count of keys that have already been used and a mutable reference to a matrix (or vector) that tracks which keys have been used.

platform_version: The platform version which determines the structure of the identity key.

Returns

Result<Self, ProtocolError>: If successful, returns an instance of Self. In case of an error, it returns a ProtocolError.

Errors

ProtocolError::PublicKeyGenerationError: This error is returned if too many keys have already been created.

ProtocolError::UnknownVersionMismatch: This error is returned if the provided platform version is not recognized.

pub fn random_authentication_key_with_private_key( id: u32, seed: Option<u64>, platform_version: &PlatformVersion, ) -> Result<(IdentityPublicKey, [u8; 32]), ProtocolError>

Generates a random authentication key and its corresponding private key based on the platform version.

Parameters

id: The KeyID for the generated key.

seed: A seed that will create a random number generator StdRng.

used_key_matrix: An optional tuple that contains the count of keys that have already been used and a mutable reference to a matrix (or vector) that tracks which keys have been used.

platform_version: The platform version which determines the structure of the identity key.

Returns

Result<(Self, Vec<u8>), ProtocolError>: If successful, returns an instance of Self and the private key as Vec<u8>. In case of an error, it returns a ProtocolError.

Errors

ProtocolError::PublicKeyGenerationError: This error is returned if too many keys have already been created.

ProtocolError::UnknownVersionMismatch: This error is returned if the provided platform version is not recognized.

pub fn random_authentication_key_with_private_key_with_rng( id: u32, rng: &mut StdRng, used_key_matrix: Option<(u32, &mut Vec<bool>)>, platform_version: &PlatformVersion, ) -> Result<(IdentityPublicKey, [u8; 32]), ProtocolError>

Generates a random authentication key and its corresponding private key based on the platform version.

Parameters

id: The KeyID for the generated key.

rng: A mutable reference to a random number generator of type StdRng.

used_key_matrix: An optional tuple that contains the count of keys that have already been used and a mutable reference to a matrix (or vector) that tracks which keys have been used.

platform_version: The platform version which determines the structure of the identity key.

Returns

Result<(Self, Vec<u8>), ProtocolError>: If successful, returns an instance of Self and the private key as Vec<u8>. In case of an error, it returns a ProtocolError.

Errors

ProtocolError::PublicKeyGenerationError: This error is returned if too many keys have already been created.

ProtocolError::UnknownVersionMismatch: This error is returned if the provided platform version is not recognized.

pub fn random_key_with_rng( id: u32, rng: &mut StdRng, used_key_matrix: Option<(u32, &mut Vec<bool>)>, platform_version: &PlatformVersion, ) -> Result<IdentityPublicKey, ProtocolError>

Generates a random key based on the platform version.

Parameters

id: The KeyID for the generated key.

rng: A mutable reference to a random number generator of type StdRng.

used_key_matrix: An optional tuple that contains the count of keys that have already been used and a mutable reference to a matrix (or vector) that tracks which keys have been used.

platform_version: The platform version which determines the structure of the identity key.

Returns

Result<Self, ProtocolError>: If successful, returns an instance of Self. In case of an error, it returns a ProtocolError.

Errors

ProtocolError::PublicKeyGenerationError: This error is returned if too many keys have already been created.

ProtocolError::UnknownVersionMismatch: This error is returned if the provided platform version is not recognized.

pub fn random_key_with_known_attributes( id: u32, rng: &mut StdRng, purpose: Purpose, security_level: SecurityLevel, key_type: KeyType, contract_bounds: Option<ContractBounds>, platform_version: &PlatformVersion, ) -> Result<(IdentityPublicKey, [u8; 32]), ProtocolError>

Generates a random key based on the platform version.

Parameters

id: The KeyID for the generated key.

rng: A mutable reference to a random number generator of type StdRng.

used_key_matrix: An optional tuple that contains the count of keys that have already been used and a mutable reference to a matrix (or vector) that tracks which keys have been used.

platform_version: The platform version which determines the structure of the identity key.

Returns

Result<Self, ProtocolError>: If successful, returns an instance of Self. In case of an error, it returns a ProtocolError.

Errors

ProtocolError::PublicKeyGenerationError: This error is returned if too many keys have already been created.

ProtocolError::UnknownVersionMismatch: This error is returned if the provided platform version is not recognized.

pub fn random_ecdsa_master_authentication_key_with_rng( id: u32, rng: &mut StdRng, platform_version: &PlatformVersion, ) -> Result<(IdentityPublicKey, [u8; 32]), ProtocolError>

Generates a random ECDSA master authentication public key along with its corresponding private key.

This method constructs a random ECDSA (using the secp256k1 curve) master authentication public key and returns both the public key and its corresponding private key.

Parameters

id: The KeyID for the generated key.

rng: A mutable reference to the random number generator.

Returns

(Self, Vec<u8>): A tuple where the first element is an instance of the IdentityPublicKey struct, and the second element is the corresponding private key.

pub fn random_voting_key_with_rng( id: u32, rng: &mut StdRng, platform_version: &PlatformVersion, ) -> Result<(IdentityPublicKey, [u8; 32]), ProtocolError>

pub fn random_ecdsa_master_authentication_key( id: u32, seed: Option<u64>, platform_version: &PlatformVersion, ) -> Result<(IdentityPublicKey, [u8; 32]), ProtocolError>

Generates a random ECDSA master-level authentication public key along with its corresponding private key.

This method constructs a random ECDSA (using the secp256k1 curve) high-level authentication public key and returns both the public key and its corresponding private key.

Parameters

id: The KeyID for the generated key.

seed: A seed that will create a random number generator StdRng.

Returns

(Self, Vec<u8>): A tuple where the first element is an instance of the IdentityPublicKey struct, and the second element is the corresponding private key.

pub fn random_ecdsa_critical_level_authentication_key( id: u32, seed: Option<u64>, platform_version: &PlatformVersion, ) -> Result<(IdentityPublicKey, [u8; 32]), ProtocolError>

Generates a random ECDSA critical-level authentication public key along with its corresponding private key.

This method constructs a random ECDSA (using the secp256k1 curve) high-level authentication public key and returns both the public key and its corresponding private key.

Parameters

id: The KeyID for the generated key.

seed: A seed that will create a random number generator StdRng.

Returns

(Self, Vec<u8>): A tuple where the first element is an instance of the IdentityPublicKey struct, and the second element is the corresponding private key.

pub fn random_ecdsa_critical_level_authentication_key_with_rng( id: u32, rng: &mut StdRng, platform_version: &PlatformVersion, ) -> Result<(IdentityPublicKey, [u8; 32]), ProtocolError>

Generates a random ECDSA high-level authentication public key along with its corresponding private key.

This method constructs a random ECDSA (using the secp256k1 curve) high-level authentication public key and returns both the public key and its corresponding private key.

Parameters

id: The KeyID for the generated key.

rng: A mutable reference to the random number generator.

Returns

(Self, Vec<u8>): A tuple where the first element is an instance of the IdentityPublicKey struct, and the second element is the corresponding private key.

pub fn random_masternode_owner_key( id: u32, seed: Option<u64>, platform_version: &PlatformVersion, ) -> Result<(IdentityPublicKey, [u8; 32]), ProtocolError>

Generates a random ECDSA critical-level authentication key for a masternode owner.

This function generates a random key that can be used for owner authentication in a masternode context. The function accepts an optional seed for deterministic key generation, or uses entropy-based randomness if no seed is provided.

Parameters

id: The identifier (KeyID) for the masternode owner key.

seed: An optional u64 value used to seed the random number generator. If None, the RNG will be seeded from entropy.

platform_version: A reference to the PlatformVersion struct, which is used to determine the correct key structure version.

Returns

Returns a tuple containing the generated IdentityPublicKey for the masternode owner and the corresponding private key as a byte vector.

Errors

Returns a ProtocolError if the platform version is not supported.

pub fn random_masternode_owner_key_with_rng( id: u32, rng: &mut StdRng, platform_version: &PlatformVersion, ) -> Result<(IdentityPublicKey, [u8; 32]), ProtocolError>

Generates a random ECDSA critical-level authentication key for a masternode owner using a custom RNG.

This function generates a random key using a given random number generator (RNG). This is useful when specific control over the randomness is needed.

Parameters

id: The identifier (KeyID) for the masternode owner key.

rng: A mutable reference to a StdRng instance used to generate randomness.

platform_version: A reference to the PlatformVersion struct, which is used to determine the correct key structure version.

Returns

Returns a tuple containing the generated IdentityPublicKey for the masternode owner and the corresponding private key as a byte vector.

Errors

Returns a ProtocolError if the platform version is not supported.

pub fn random_masternode_transfer_key( id: u32, seed: Option<u64>, platform_version: &PlatformVersion, ) -> Result<(IdentityPublicKey, [u8; 32]), ProtocolError>

Generates a random ECDSA critical-level transfer key for a masternode.

This function generates a random key for use in transferring ownership of a masternode. An optional seed can be provided for deterministic key generation, or entropy-based randomness is used if no seed is given.

Parameters

id: The identifier (KeyID) for the masternode transfer key.

seed: An optional u64 value used to seed the random number generator. If None, the RNG will be seeded from entropy.

platform_version: A reference to the PlatformVersion struct, which is used to determine the correct key structure version.

Returns

Returns a tuple containing the generated IdentityPublicKey for the masternode transfer key and the corresponding private key as a byte vector.

Errors

Returns a ProtocolError if the platform version is not supported.

pub fn random_masternode_transfer_key_with_rng( id: u32, rng: &mut StdRng, platform_version: &PlatformVersion, ) -> Result<(IdentityPublicKey, [u8; 32]), ProtocolError>

Generates a random ECDSA critical-level transfer key for a masternode using a custom RNG.

This function generates a random key for masternode transfers using a given random number generator (RNG).

Parameters

id: The identifier (KeyID) for the masternode transfer key.

rng: A mutable reference to a StdRng instance used to generate randomness.

platform_version: A reference to the PlatformVersion struct, which is used to determine the correct key structure version.

Returns

Returns a tuple containing the generated IdentityPublicKey for the masternode transfer key and the corresponding private key as a byte vector.

Errors

Returns a ProtocolError if the platform version is not supported.

pub fn random_ecdsa_high_level_authentication_key( id: u32, seed: Option<u64>, platform_version: &PlatformVersion, ) -> Result<(IdentityPublicKey, [u8; 32]), ProtocolError>

Generates a random ECDSA high-level authentication public key along with its corresponding private key.

This method constructs a random ECDSA (using the secp256k1 curve) high-level authentication public key and returns both the public key and its corresponding private key.

Parameters

id: The KeyID for the generated key.

seed: A seed that will create a random number generator StdRng.

Returns

(Self, Vec<u8>): A tuple where the first element is an instance of the IdentityPublicKey struct, and the second element is the corresponding private key.

pub fn random_ecdsa_high_level_authentication_key_with_rng( id: u32, rng: &mut StdRng, platform_version: &PlatformVersion, ) -> Result<(IdentityPublicKey, [u8; 32]), ProtocolError>

Generates a random ECDSA high-level authentication public key along with its corresponding private key.

This method constructs a random ECDSA (using the secp256k1 curve) high-level authentication public key and returns both the public key and its corresponding private key.

Parameters

id: The KeyID for the generated key.

rng: A mutable reference to the random number generator.

Returns

(Self, Vec<u8>): A tuple where the first element is an instance of the IdentityPublicKey struct, and the second element is the corresponding private key.

pub fn random_authentication_keys_with_rng( key_count: u32, rng: &mut StdRng, platform_version: &PlatformVersion, ) -> Result<Vec<IdentityPublicKey>, ProtocolError>

pub fn random_authentication_keys_with_private_keys_with_rng( start_id: u32, key_count: u32, rng: &mut StdRng, platform_version: &PlatformVersion, ) -> Result<Vec<(IdentityPublicKey, [u8; 32])>, ProtocolError>

pub fn main_keys_with_random_authentication_keys_with_private_keys_with_rng( key_count: u32, rng: &mut StdRng, platform_version: &PlatformVersion, ) -> Result<Vec<(IdentityPublicKey, [u8; 32])>, ProtocolError>

pub fn random_unique_keys_with_rng( key_count: u32, rng: &mut StdRng, platform_version: &PlatformVersion, ) -> Result<Vec<IdentityPublicKey>, ProtocolError>

pub fn random_keys_with_rng( key_count: u32, rng: &mut StdRng, platform_version: &PlatformVersion, ) -> Vec<IdentityPublicKey>

impl IdentityPublicKey

pub fn is_master(&self) -> bool

Checks if public key security level is MASTER

pub fn max_possible_size_key( id: u32, platform_version: &PlatformVersion, ) -> Result<IdentityPublicKey, ProtocolError>

Generates an identity public key with the maximum possible size based on the platform version.

This method constructs a key of the largest possible size for the given platform version. This can be useful for stress testing or benchmarking purposes.

Parameters

id: The KeyID for the generated key.

platform_version: The platform version which determines the structure of the identity key.

Returns

Self: An instance of the IdentityPublicKey struct.

pub fn default_versioned( platform_version: &PlatformVersion, ) -> Result<IdentityPublicKey, ProtocolError>